I-PV: interactive protein sequence/feature viewer
First, head on to http://i-pv.org/ipv_minimal.html to have an idea about
what I-PV is and what it can do for you. This folder contains the
necessary files to run I-PV, a circos module for interactive sequence
visualization/viewer. Download the folder into any location on your
disk, then specify the correct path on SNPtoAA._43pl script at line 870
to the version of circos you want to use.  Run the script and follow the
instructions.